requirements.txt
Pillow
Flask
tensorflow>=2.0.0
mysql-connector-python
Flask-MySQLdb